LPS-56703 CustomAttributesDisplay
